requirements.txt
easyocr==1.7.1
filelock==3.14.0
fsspec==2024.5.0
imageio==2.34.1
intel-openmp==2021.4.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
pyclipper==1.3.0.post5
python-bidi==0.4.2
PyYAML==6.0.1
scikit-image==0.23.2
scipy==1.13.1
shapely==2.0.4
six==1.16.0
sympy==1.12
tbb==2021.12.0
tifffile==2024.5.22
torch==2.3.0
torchvision==0.18.0
typing_extensions==4.12.0
